Move types and prompts
