+++ packages/insomnia-inso/src/db/models/__tests__/environment.test.js
@@ -1,8 +1,8 @@
 // @flow
 import type { Database } from '../../index';
 import { emptyDb } from '../../index';
-import type { Environment, Workspace } from '../../types';
-import { generateIdIsh } from '../../prompts';
+import type { Environment, Workspace } from '../types';
+import { generateIdIsh } from '../utils';
 import enquirer from 'enquirer';
 import { loadEnvironment, promptEnvironment } from '../environment';
