+++ packages/insomnia-inso/src/db/models/__tests__/workspace.test.js
@@ -1,8 +1,8 @@
 // @flow
 import type { Database } from '../../index';
 import { emptyDb } from '../../index';
-import type { Workspace } from '../../types';
-import { generateIdIsh } from '../../prompts';
+import type { Workspace } from '../types';
+import { generateIdIsh } from '../utils';
 import { loadWorkspace } from '../workspace';
 
 describe('workspace', () => {
