+++ packages/insomnia-inso/src/db/models/workspace.js
@@ -1,8 +1,8 @@
 // @flow
 import type { Database } from '../index';
-import type { Workspace } from '../types';
+import type { Workspace } from './types';
 import { mustFindSingleOrNone } from '../index';
-import { matchIdIsh } from '../prompts';
+import { matchIdIsh } from './utils';
 
 export const loadWorkspace = (db: Database, identifier: string): ?Workspace =>
   mustFindSingleOrNone(db.Workspace, s => matchIdIsh(s, identifier) || s.name === identifier);
